+++ src/shared/newGame/types.ts
